Added a check for logged in users
Checks if the account's name has a value > 0 (logged in) in the database and returns an error if it is logged in and proceeds without error if it is not logged in. 

This is helpful as many servers currently cannot give vote points/rewards to users who are logged in.